# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: DealFlow360 API
  description: Comprehensive quotation-to-cash platform API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /quotes/{quoteId}/recommendations:
    get:
      summary: Get upsell recommendations
      tags:
        - Recommendations
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendations list
  /quotes/{quoteId}/recommendations/{productId}/dismiss:
    post:
      summary: Dismiss recommendation
      tags:
        - Recommendations
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendation dismissed
  /quotes/{quoteId}/recommendations/{productId}/add:
    post:
      summary: Add recommended product
      tags:
        - Recommendations
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: string
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product added
...
  /pricing/price-lists:
    get:
      summary: List price lists
      tags:
        - Pricing
      responses:
        '200':
          description: Price list
    post:
      summary: Create price list
      tags:
        - Pricing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                currency:
                  type: string
                effectiveFrom:
                  type: string
                  format: date-time
                effectiveTo:
                  type: string
                  format: date-time
                priority:
                  type: number
      responses:
        '201':
          description: Price list created
  /pricing/price-lists/{priceListId}:
    patch:
      summary: Update price list
      tags:
        - Pricing
      parameters:
        - name: priceListId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                active:
                  type: boolean
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Price list updated
  /pricing/price-lists/{priceListId}/rules:
    get:
      summary: List price rules
      tags:
        - Pricing
      parameters:
        - name: priceListId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Price rules
    post:
      summary: Create price rule
      tags:
        - Pricing
      parameters:
        - name: priceListId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                categoryId:
                  type: string
                tierId:
                  type: string
                minQuantity:
                  type: string
                unitPrice:
                  type: string
                priority:
                  type: number
                effectiveFrom:
                  type: string
                  format: date-time
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Price rule created
  /pricing/price-rules/{ruleId}:
    patch:
      summary: Update price rule
      tags:
        - Pricing
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                unitPrice:
                  type: string
                priority:
                  type: number
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Price rule updated
    delete:
      summary: Delete price rule
      tags:
        - Pricing
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Price rule deleted
  /pricing/discount-limits:
    get:
      summary: List discount limits
      tags:
        - Pricing
      responses:
        '200':
          description: Discount limits
    post:
      summary: Create discount limit
      tags:
        - Pricing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                tierId:
                  type: string
                categoryId:
                  type: string
                productId:
                  type: string
                maxDiscountPct:
                  type: string
                priority:
                  type: number
                effectiveFrom:
                  type: string
                  format: date-time
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Discount limit created
  /pricing/discount-limits/{limitId}:
    patch:
      summary: Update discount limit
      tags:
        - Pricing
      parameters:
        - name: limitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                maxDiscountPct:
                  type: string
                active:
                  type: boolean
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Discount limit updated
  /pricing/taxes:
    get:
      summary: List taxes
      tags:
        - Pricing
      responses:
        '200':
          description: Tax list
    post:
      summary: Create tax
      tags:
        - Pricing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                rate:
                  type: string
                behavior:
                  type: string
                  enum: [INCLUSIVE, EXCLUSIVE]
                effectiveFrom:
                  type: string
                  format: date-time
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Tax created
  /pricing/taxes/{taxId}:
    patch:
      summary: Update tax
      tags:
        - Pricing
      parameters:
        - name: taxId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rate:
                  type: string
                active:
                  type: boolean
                effectiveTo:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Tax updated
  /pricing/subscription-plans:
    get:
      summary: List subscription plans
      tags:
        - Pricing
      responses:
        '200':
          description: Subscription plans
    post:
      summary: Create subscription plan
      tags:
        - Pricing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                code:
                  type: string
                interval:
                  type: string
                  enum: [DAY, WEEK, MONTH, YEAR]
                intervalCount:
                  type: number
                prorationConvention:
                  type: string
                  enum: [CALENDAR_DAYS, THIRTY_DAY_MONTH]
                cancellationRules:
                  type: object
      responses:
        '201':
          description: Subscription plan created
  /pricing/subscription-plans/{planId}:
    patch:
      summary: Update subscription plan
      tags:
        - Pricing
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                active:
                  type: boolean
                cancellationRules:
                  type: object
      responses:
        '200':
          description: Subscription plan updated
...
  /reporting/exports:
    get:
      summary: List export jobs
      tags:
        - Reporting
      responses:
        '200':
          description: Export jobs list
    post:
      summary: Create export job
      tags:
        - Reporting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportType:
                  type: string
                  enum: [QUOTES, ORDERS, INVOICES, CUSTOMERS, INVENTORY]
                format:
                  type: string
                  enum: [CSV, XLSX, PDF]
                filters:
                  type: object
      responses:
        '201':
          description: Export job created
...
  /reporting/exports/{jobId}/download:
    get:
      summary: Download export
      tags:
        - Reporting
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Download URL
        '202':
          description: Export still processing
  /reporting/exports/{jobId}:
    delete:
      summary: Delete export job
      tags:
        - Reporting
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Export job deleted
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session